wake-lock/wakelock-onactivechange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Wake Lock 'onactivechange' Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

promise_test(t => {
  let request, wakelock, eventWatcher;
  return navigator.getWakeLock("screen")
  .then(wakeLock => {
    wakelock = wakeLock;
    assert_false(wakelock.active, "the active is false before wake lock is acquired");
    eventWatcher = new EventWatcher(t, wakelock, 'activechange');
    request = wakelock.createRequest();
    return eventWatcher.wait_for('activechange')
  })
  .then(evt => {
    assert_true(evt.isTrusted && !evt.bubbles && !evt.cancelable && evt instanceof Event, 'a simple event is fired');
    assert_equals(evt.type, "activechange", 'the event name is "activechange"');
    assert_equals(evt.target, wakelock, 'event.target is WakeLock.');
    assert_true(wakelock.active, "the active is true when wake lock is acquired");
    request.cancel();
    return eventWatcher.wait_for('activechange')
  })
  .then(() => {
    assert_false(wakelock.active, "the active is false when wake lock is released");
  });
}, "Test that 'activechange' event is fire and WakeLock.active is valid");

</script>
